+++ dashboard/src/components/TableList/index.js
@@ -1,0 +1,5 @@
+export { default } from "./TableList";
+export { default as TableListHeader } from "./TableListHeader";
+export { default as TableListItem } from "./TableListItem";
+export { default as TableListSelect } from "./TableListSelect";
+export { default as TableListEmptyState } from "./TableListEmptyState";