+++ dashboard/src/pages/EventsPage.js
@@ -2,8 +2,9 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import { graphql } from "react-relay";
-import Typography from "material-ui/Typography";
+import { routerShape, matchShape } from "found";
 import { withStyles } from "material-ui/styles";
+import Typography from "material-ui/Typography";
 
 import AppContent from "../components/AppContent";
 import EventsContainer from "../components/EventsContainer";
@@ -26,32 +27,47 @@ const styles = {
 
 class EventsPage extends React.Component {
   static propTypes = {
-    // eslint-disable-next-line react/forbid-prop-types
     classes: PropTypes.object.isRequired,
+    location: PropTypes.shape({
+      query: PropTypes.object.isRequired,
+    }).isRequired,
+    router: routerShape.isRequired,
+    match: matchShape.isRequired,
   };
 
   static query = graphql`
     query EventsPageQuery(
       $filter: String
+      $order: EventsListOrder = SEVERITY
       $environment: String!
       $organization: String!
     ) {
-      viewer {
-        ...EventsContainer_viewer
-      }
       environment(organization: $organization, environment: $environment) {
         ...EventsContainer_environment
       }
     }
   `;
 
-  state = { inputValue: "" };
+  constructor(props) {
+    super(props);
+    this.state = {
+      filterValue: props.location.query.filter,
+    };
+  }
 
-  requerySearchBox = query => {
-    this.setState({ inputValue: query });
-    // TODO return to this and make it actually query
-    // eslint-disable-next-line no-console
-    console.info("query", query);
+  changeQuery = (key, val) => {
+    const { match, router } = this.props;
+    const query = new URLSearchParams(match.location.query);
+
+    query.set(key, val);
+    router.push(`${match.location.pathname}?${query.toString()}`);
+  };
+
+  requerySearchBox = filterValue => {
+    this.setState({ filterValue });
+    if (filterValue.length > 10) {
+      this.changeQuery("filter", filterValue);
+    }
   };
 
   render() {
@@ -65,7 +81,7 @@ class EventsPage extends React.Component {
             </Typography>
             <SearchBox
               onUpdateInput={this.requerySearchBox}
-              state={this.state.inputValue}
+              state={this.state.filterValue}
             />
           </div>
           <div className={classes.container}>